Ensurre worker waits for migrations completion
In some case when using docker compose up, the worker processes may
start before the web process's database migration is completed.
This trigger byzantine errors that are not obvious except by looking in
the logs and will let you start a pipeline that will be queued forever.
To solve this issue we install the wait-for-it script available in
Debian by @vishnubob and as suggested in the Docker documentation.
In the docker-compose.yml, we let the worker wait for the web processing
to be complete when gunicorn exposes port 8000 and web container is
available.

Reported-by: Lucas Gonze @lucasgonze
Reference: https://github.com/nexB/scancode.io/issues/387
Reference: https://docs.docker.com/compose/startup-order/
Reference: https://github.com/vishnubob/wait-for-it
Reference: https://tracker.debian.org/pkg/wait-for-it
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>