Do not fail project on memory spikes
We have an unresolved issued of multi-gigabyte memory spikes when
scanning files for licenses in some rare cases, like it is reported
at https://github.com/aboutcode-org/scancode-toolkit/issues/3711
Currently we fail the project on encountering such memory spikes,
but we don't get any scan info from the rest of the project because
we raise the Exception. This adds a project error instead of raising
an Exception, so we can continue scanning the rest of the resources.

Reference: https://github.com/aboutcode-org/scancode-toolkit/issues/3711
Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>