Add rust binary support #1435 (#1488)
* Add rust binary support for packages and symbols

* Add support to get packages from rust binaries
* Add support for rust source to binary mapping using symbols

Reference: https://github.com/aboutcode-org/scancode.io/issues/1435
Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>

* Add support for single binary file on the to/ side

We were failing on the d2d pipelines if we uploaded a single binary
on the to side, as the d2d pipeline requires uploading an archive only.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>

* Add warning instead of logger line in go/elf map

We were having a lot of extra logger lines in certain cases previously
on selecting wrong options, as the traceback was being shown in the log
too. So this creates a warning instead of a log entry.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>

* Address feedback

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>

* Update CHANGELOG and documention

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>

---------

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>