Close existing connection before job execution in worker (#1959)
* Close existing DB connection before job execution in worker

The tech stack of django-rq seems to have issues with TLS connections getting corrupted when using the Worker class, which uses fork. This is relevant when an external database is used. This problem can be avoided by closing the existing connection prior to the call to `execute_job`. For reference see: https://github.com/aboutcode-org/scancode.io/issues/1523 and https://github.com/rq/django-rq/issues/650

Signed-off-by: Robert Guetzkow <robert.guetzkow@betasystems.com>

* Move ScanCodeIOWorker modification for TLS connection issues to dedicated class

The current default is kept and the new worker class with the workaround can be enabled through the `--worker-class` option of rqworker, if needed

Signed-off-by: Robert Guetzkow <robert.guetzkow@betasystems.com>

---------

Signed-off-by: Robert Guetzkow <robert.guetzkow@betasystems.com>