Add support for authors in ORT package list generation (#1988)
* Add support for authors in ORT package list generation

This updates the ORT pipe to include the 'authors' field in the generated package list YAML. The authors are extracted from the package parties where the role is 'maintainer'.

Ref: #1987
Signed-off-by: Fabian Noll <Fabian.Noll@googlemail.com>

* Include 'author' role in ORT authors list and update test data

This extends the authors extraction to include parties with the 'author' role, in addition to 'maintainer'. It also updates the asgiref test data to match the expected output.

Ref: #1987
Signed-off-by: Fabian Noll <Fabian.Noll@googlemail.com>

---------

Signed-off-by: Fabian Noll <Fabian.Noll@googlemail.com>